Add accept and reject modes, refactor to use individual files
