+++ tests/cookie-management/rejectCookies.test.ts
@@ -53,6 +53,66 @@ describe('Reject Cookies tests', () => {
     expect(document.cookie).to.equal(defaultRejectedCookie)
   })
 
+  it('should create a cookie and hide all banners when using reject-only mode', () => {
+    const html = `
+    <html lang="en">
+      <body>
+        <div id="cookie-banner">
+          <div id="accept-or-reject-message"></div>
+          <div id="accepted-cookies-message" hidden></div>
+          <div id="rejected-cookies-message" hidden></div>
+        </div>
+      </body>
+    </html>`
+    createJSDOM(html)
+    const callback = spy()
+
+    rejectCookies(callback, 'reject-only')
+
+    const mockAcceptOrRejectMessage = document.getElementById('accept-or-reject-message')
+    const mockCookieBanner = document.getElementById('cookie-banner')
+    const mockCookiesAcceptedMessage = document.getElementById('accepted-cookies-message')
+    const mockCookiesRejectedMessage = document.getElementById('rejected-cookies-message')
+
+    expect(mockAcceptOrRejectMessage).to.have.attribute('hidden')
+    expect(mockCookieBanner).to.have.attribute('hidden')
+    expect(mockCookiesAcceptedMessage).to.have.attribute('hidden')
+    expect(mockCookiesRejectedMessage).to.have.attribute('hidden')
+    expect(document.cookie).to.equal(defaultRejectedCookie)
+  })
+
+  it('should create a cookie and log an error if an incorrect mode is passed in', () => {
+    const html = `
+    <html lang="en">
+      <body>
+        <div id="cookie-banner">
+          <div id="accept-or-reject-message"></div>
+          <div id="accepted-cookies-message" hidden></div>
+          <div id="rejected-cookies-message" hidden></div>
+        </div>
+      </body>
+    </html>`
+    createJSDOM(html)
+    const consoleSpy = spy()
+    window.console.error = consoleSpy
+    const callback = spy()
+
+    // @ts-expect-error
+    rejectCookies(callback, 'error-mode')
+
+    const mockAcceptOrRejectMessage = document.getElementById('accept-or-reject-message')
+    const mockCookieBanner = document.getElementById('cookie-banner')
+    const mockCookiesAcceptedMessage = document.getElementById('accepted-cookies-message')
+    const mockCookiesRejectedMessage = document.getElementById('rejected-cookies-message')
+
+    expect(mockAcceptOrRejectMessage).not.to.have.attribute('hidden')
+    expect(mockCookieBanner).not.to.have.attribute('hidden')
+    expect(mockCookiesAcceptedMessage).to.have.attribute('hidden')
+    expect(mockCookiesRejectedMessage).to.have.attribute('hidden')
+    expect(consoleSpy).to.have.been.called()
+    expect(document.cookie).to.equal(defaultRejectedCookie)
+  })
+
   it('should set a cookie, hide the accept or reject message, show the rejected message and log an error if the callback throws', () => {
     const html = `
     <html lang="en">
